updating

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,13 +8,13 @@
 
 int main(void)
 {
-int i, n1, sum, tem;
+long int i, n1, sum, tem;
 
 n1 = 1, sum = 0;
 
 for (i = 1 ; i < 50 ; i++)
 {
-printf("%d", sum + n1);
+printf("%ld", sum + n1);
 tem = n1;
 n1 = sum + n1;
 sum = tem;
